eutonafila/deploy_package/requirements/base.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
channels==4.0.0
daphne==4.0.0
whitenoise==6.5.0
python-dotenv==1.0.0
gunicorn==21.2.0